requirements.txt
Django
djangorestframework
Celery
clickhouse-driver
psycopg2
psycopg2-binary
